+++ src/main/java/de/codeinfection/quickwango/AntiGuest/FilteredItemPrevention.java
@@ -4,23 +4,40 @@ import java.util.HashSet;
 import org.bukkit.Material;
 import org.bukkit.Server;
 import org.bukkit.configuration.ConfigurationSection;
+import org.bukkit.entity.Player;
+import org.bukkit.event.Cancellable;
 import org.bukkit.plugin.Plugin;
 
 /**
  * This class represents a filterable Prevention related to Materials
  *
  * @author Phillip Schichtel
+ * @TODO reduce code duplication in prevent and preventThrottled
  */
 public abstract class FilteredItemPrevention extends FilteredPrevention
 {
-    public FilteredItemPrevention(String name, Plugin plugin)
+    private final boolean ignoreBlocks;
+
+    public FilteredItemPrevention(final String name, final Plugin plugin)
+    {
+        this(name, plugin, false);
+    }
+
+    public FilteredItemPrevention(final String name, final Plugin plugin, final boolean ignoreBlocks)
     {
         super(name, plugin);
+        this.ignoreBlocks = ignoreBlocks;
     }
 
     public FilteredItemPrevention(String name, String permission, Plugin plugin)
     {
+        this(name, permission, plugin, false);
+    }
+
+    public FilteredItemPrevention(String name, String permission, Plugin plugin, final boolean ignoreBlocks)
+    {
         super(name, permission, plugin);
+        this.ignoreBlocks = ignoreBlocks;
     }
 
     /**
@@ -40,7 +57,7 @@ public abstract class FilteredItemPrevention extends FilteredPrevention
 
     /**
      * This method parses the string list to Material instances
-     * 
+     *
      * @param server a Server instance
      * @param config the config of this prevention
      */
@@ -54,11 +71,68 @@ public abstract class FilteredItemPrevention extends FilteredPrevention
         for (Object item : this.filterItems)
         {
             Material material = Material.matchMaterial(String.valueOf(item));
-            if (item != null)
+            if (material != null)
             {
                 newItems.add(material);
             }
         }
         this.filterItems = newItems;
+    }
+
+    /**
+     * This method checks whether the player can do the subaction
+     *
+     * @param player the player
+     * @param material the material
+     * @return true if he can
+     */
+    public boolean can(final Player player, final Material material)
+    {
+        if (this.ignoreBlocks && material.getId() <= 256)
+        {
+            return true;
+        }
+        else
+        {
+            return super.can(player, material);
+        }
+    }
+
+    /**
+     * Prevents the action if the player can't pass it
+     *
+     * @param event a cancellable event
+     * @param player the player
+     * @param material the material
+     * @return true if the action was prevented
+     */
+    public boolean prevent(final Cancellable event, final Player player, final Material material)
+    {
+        if (!this.can(player, material))
+        {
+            event.setCancelled(true);
+            this.sendMessage(player);
+            return true;
+        }
+        return false;
+    }
+
+    /**
+     * The same as prevent(Cancellable, Player, Object) except that it is throttled
+     *
+     * @param event a cancellable event
+     * @param player the player
+     * @param material the item representing the subaction
+     * @return true if the action was prevented
+     */
+    public boolean preventThrottled(final Cancellable event, final Player player, final Material material)
+    {
+        if (!this.can(player, material))
+        {
+            event.setCancelled(true);
+            this.sendThrottledMessage(player);
+            return true;
+        }
+        return false;
     }
 }